moved measurement unit into its own file
